Extend task to read output path from the config file. Add validations.
