+++ Directory.Build.props
@@ -1,4 +1,4 @@
-﻿<Project>
+<Project>
     <PropertyGroup>
         <Authors>Stanislav Silin</Authors>
         <PackageProjectUrl>https://github.com/byme8/ZeroQL</PackageProjectUrl>
