+++ ZeroQL.sln
@@ -32,6 +32,12 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZeroQL.Runtime", "src\ZeroQ
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestStandardLibrary", "src\TestApp\TestStandardLibrary\TestStandardLibrary.csproj", "{2B4ADDAC-5AE6-4A57-B8D7-EA16FF028BA1}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZeroQL.BuildTaskTest", "src\TestApp\ZeroQL.BuildTaskTest\ZeroQL.BuildTaskTest.csproj", "{19CE1308-62C8-487E-B248-2E16C63AFC87}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZeroQL.MSBuild", "src\ZeroQL.MSBuild\ZeroQL.MSBuild.csproj", "{91BA7B37-430B-4953-B97E-748C730637B3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZeroQL.Core", "src\ZeroQL.Core\ZeroQL.Core.csproj", "{271B3BFF-A3D4-4296-883A-DFFA050C5F89}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -80,6 +86,18 @@ Global
 		{2B4ADDAC-5AE6-4A57-B8D7-EA16FF028BA1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2B4ADDAC-5AE6-4A57-B8D7-EA16FF028BA1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2B4ADDAC-5AE6-4A57-B8D7-EA16FF028BA1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{19CE1308-62C8-487E-B248-2E16C63AFC87}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{19CE1308-62C8-487E-B248-2E16C63AFC87}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{19CE1308-62C8-487E-B248-2E16C63AFC87}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{19CE1308-62C8-487E-B248-2E16C63AFC87}.Release|Any CPU.Build.0 = Release|Any CPU
+		{91BA7B37-430B-4953-B97E-748C730637B3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{91BA7B37-430B-4953-B97E-748C730637B3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{91BA7B37-430B-4953-B97E-748C730637B3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{91BA7B37-430B-4953-B97E-748C730637B3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{271B3BFF-A3D4-4296-883A-DFFA050C5F89}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{271B3BFF-A3D4-4296-883A-DFFA050C5F89}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{271B3BFF-A3D4-4296-883A-DFFA050C5F89}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{271B3BFF-A3D4-4296-883A-DFFA050C5F89}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{09D4E53A-C80D-430F-AB2C-CBC85FE8C98F} = {B240CB61-7286-495B-8679-024640C040DA}
@@ -94,5 +112,8 @@ Global
 		{A19EA773-661E-4E4B-9D66-84A47DE8F87F} = {7392465B-92A7-4702-B2D5-C2D29FDB3B3D}
 		{7FC88940-CF09-4FB5-8724-015538C64E8A} = {B240CB61-7286-495B-8679-024640C040DA}
 		{2B4ADDAC-5AE6-4A57-B8D7-EA16FF028BA1} = {1088FDB6-BE3D-4D02-B744-2BF4CA530980}
+		{19CE1308-62C8-487E-B248-2E16C63AFC87} = {1088FDB6-BE3D-4D02-B744-2BF4CA530980}
+		{91BA7B37-430B-4953-B97E-748C730637B3} = {B240CB61-7286-495B-8679-024640C040DA}
+		{271B3BFF-A3D4-4296-883A-DFFA050C5F89} = {B240CB61-7286-495B-8679-024640C040DA}
 	EndGlobalSection
 EndGlobal
