+++ src/TestApp/ZeroQL.BuildTaskTest/Program.cs
@@ -1,0 +1,3 @@
+﻿// See https://aka.ms/new-console-template for more information
+
+Console.WriteLine("Hello, World!");
