+++ src/TestApp/ZeroQL.BuildTaskTest/ZeroQL.BuildTaskTest.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net7.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+<!--        <ZeroQLCLIValidation>False</ZeroQLCLIValidation>-->
+    </PropertyGroup>
+
+    <ItemGroup>
+<!--      <PackageReference Include="ZeroQL" Version="999.2308.1622.4513" />-->
+    </ItemGroup>
+    
+</Project>
