+++ src/TestApp/ZeroQL.TestApp/ZeroQL.TestApp.csproj
@@ -7,11 +7,13 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\..\ZeroQL.Runtime\ZeroQL.Runtime.csproj" />
-        <ProjectReference Include="..\..\ZeroQL.SourceGenerators\ZeroQL.SourceGenerators.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
+        <ProjectReference Include="..\..\ZeroQL.Runtime\ZeroQL.Runtime.csproj"/>
+        <ProjectReference Include="..\..\ZeroQL.SourceGenerators\ZeroQL.SourceGenerators.csproj"
+                          OutputItemType="Analyzer"
+                          ReferenceOutputAssembly="false"/>
     </ItemGroup>
 
     <ItemGroup>
-      <Folder Include="Generated" />
+        <Folder Include="Generated"/>
     </ItemGroup>
 </Project>
