+++ src/ZeroQL.Core/Config/ZeroQLFileConfig.cs
@@ -1,11 +1,9 @@
-using System.Collections.Generic;
-using System.Linq;
 using Newtonsoft.Json;
-using ZeroQL.Internal.Enums;
+using ZeroQL.Core.Enums;
 
 #pragma warning disable CS8618
 
-namespace ZeroQL.CLI;
+namespace ZeroQL.Core.Config;
 
 public class ZeroQLFileConfig
 {
