+++ src/ZeroQL.MSBuild/ZeroQL.MSBuild.csproj
@@ -1,0 +1,18 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>netstandard2.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <LangVersion>latest</LangVersion>
+    </PropertyGroup>
+    
+    <ItemGroup>
+        <PackageReference Include="Microsoft.Build.Utilities.Core" Version="15.9.20" Pack="false" />
+    </ItemGroup>
+    
+    <ItemGroup>
+      <ProjectReference Include="..\ZeroQL.Core\ZeroQL.Core.csproj" />
+    </ItemGroup>
+
+</Project>
