+++ src/ZeroQL.SourceGenerators/ZeroQL.SourceGenerators.csproj
@@ -9,6 +9,7 @@
 
     <ItemGroup>
         <Compile Include="..\ZeroQL.Runtime\Internal\QueryKey.cs" />
+        <Compile Include="..\ZeroQL.Core\Result.cs" />
     </ItemGroup>
     
     <ItemGroup>
