+++ src/ZeroQL.Tools/Bootstrap/GraphQLGenerator.cs
@@ -7,9 +7,9 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using ZeroQL.Bootstrap.Generators;
+using ZeroQL.Core.Enums;
 using ZeroQL.Extensions;
 using ZeroQL.Internal;
-using ZeroQL.Internal.Enums;
 using ZeroQL.Schema;
 using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
