+++ src/ZeroQL.Tools/ZeroQL.Tools.csproj
@@ -9,18 +9,11 @@
 
     <ItemGroup>
         <PackageReference Include="GraphQL-Parser" Version="8.1.0" />
-        <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
-        <PackageReference Include="NJsonSchema" Version="10.9.0" />
         <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.0.0" />
-        <PackageReference Include="Microsoft.Build.Utilities.Core" Version="15.9.20" Pack="false" />
-    </ItemGroup>
-
-
-    <ItemGroup>
-        <Compile Include="..\ZeroQL.SourceGenerators\Result.cs" />
     </ItemGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\ZeroQL.Core\ZeroQL.Core.csproj" />
       <ProjectReference Include="..\ZeroQL.Runtime\ZeroQL.Runtime.csproj" />
     </ItemGroup>
